Utility\FilteredIteratorTest: split constructor specific tests to dedicated test class
